Relocate supporting files to src directory root
